requirements.txt
streamlit
tensorflow
pandas
numpy
pillow
plotly
